request.get_header removes race-condition
Context: When I use `bottle.run(..., server='cherrypy')`, I run it in multi-threaded mode. This results in two requests sharing the same environment.
They serve the old file with 200 in this case.
Fix: getting the header from the request object first, makes the actual request being used and not the request that runs in parallel.
Evaluation: This is a performance improvement.